Add question globally. Global question editing not fully working.
Editing needs separate Controller as it is too different to standard editing. Adding question globally doesn't refresh the list of questions on going back (see comment in code). Almost everything was tried.

+++ code/production/survbu_v1/survbu_v1/www/js/app_specific/questions/questions.list.controller.js
@@ -27,92 +27,69 @@
         var parentSectionId = $stateParams.parentSectionId,
             parentSurveyId = $stateParams.parentSurveyId,
             vm = angular.extend(this, {
-            parentSection: sectionsSrvc.getSectionAt(parentSectionId),
-            parentSurvey: surveysSrvc.getSurveyAt(parentSurveyId),
-            questions: questionsSrvc.getQuestions(),
-            allQuestions: questionsSrvc.returnAllQuestions(), 
-            stillWaits: questionsSrvc.isItWaiting(),
+                parentSection: sectionsSrvc.getSectionAt(parentSectionId),
+                parentSurvey: surveysSrvc.getSurveyAt(parentSurveyId),
+                questions: questionsSrvc.getQuestions(),
+                allQuestions: questionsSrvc.returnAllQuestions(), 
+                stillWaits: questionsSrvc.isItWaiting(),
 
-            stillWaiting: function () {
-                return vm.stillWaits;
-            },
-            noContent: function () {
-                return vm.questions.length === 0;
-            },
-            hideList: function () {
-                return (vm.stillWaiting() || vm.noContent());
-            },
-            hideNoItems: function () {
-                return (vm.stillWaiting() || !vm.noContent());
-            },
-            noContentAll: function () {
-                return vm.allQuestions.length === 0;
-            },
-            hideListAll: function () {
-                return (vm.stillWaiting() || vm.noContentAll());
-            },
-            hideNoItemsAll: function () {
-                return (vm.stillWaiting() || !vm.noContentAll());
-            },
-            selectDetail: function selectDetail(questionId) {
-                $state.go('questions_detail', {
-                    questionId: questionId,
-                    parentSectionId: vm.parentSection.id,
-                    parentSurveyId: vm.parentSurvey.id
-                });
-            },
-            addQuestion: function () {
-                $state.go('questions_add', {
-                    parentSectionId: vm.parentSection.id,
-                    parentSurveyId: vm.parentSurvey.id
-                });
-            },
-            addFromExisting: function () {
-                questionsSrvc.isWaiting(true);
-                $state.go('questions_addfe', {
-                    parentSectionId: vm.parentSection.id,
-                    parentSurveyId: vm.parentSurvey.id
-                });
-                questionsSrvc.updateRemainingQuestions().then(function () {
-                    $state.reload();
-                    questionsSrvc.isWaiting(false);
-                });
-            },
-            showDeleteAlert: function ($event, questionId) {
-                $event.stopPropagation();
-
-                var question = questionsSrvc.getQuestionAt(questionId);
-                $ionicPopup.confirm({
-                    title: 'Delete Question',
-                    template: 'Are you sure you want to delete \'' + question.questionText + '\'?'
-                }).then(function (response) {
-                    if (response) {
-                        var removeIndex = vm.questions.findIndex(quest => quest.id === questionId);
-                        vm.questions.splice(removeIndex, 1);
-                        vm.parentSection.questionIds.splice(removeIndex, 1);
-                        question.referenceCount = question.referenceCount - 1;
-                        questionsSrvc.updateQuestion(question).then(function () {
-                            sectionsSrvc.updateSection(vm.parentSection).then(function () {
-                                sectionsSrvc.updateSections(vm.parentSurvey.sectionIds).then(function () {
-                                    questionsSrvc.updateQuestions(vm.parentSection.questionIds);
-                                });
-                            });
-                        });
-                    } else {
-                        console.log('User pressed cancel');
-                    }
-                });
-            },
-            showDeleteAlertGlobal: function ($event, questionId) {
-                $event.stopPropagation();
-
-                var question = questionsSrvc.getQuestionAtGlobal(questionId);
-                if (question.referenceCount > 0) {
-                    $ionicPopup.alert({
-                        title: 'Can\'t delete question from global list!',
-                        template: 'Questions can only be deleted via the relevant section.'
+                stillWaiting: function () {
+                    return vm.stillWaits;
+                },
+                noContent: function () {
+                    return vm.questions.length === 0;
+                },
+                hideList: function () {
+                    return (vm.stillWaiting() || vm.noContent());
+                },
+                hideNoItems: function () {
+                    return (vm.stillWaiting() || !vm.noContent());
+                },
+                noContentAll: function () {
+                    return vm.allQuestions.length === 0;
+                },
+                hideListAll: function () {
+                    return (vm.stillWaiting() || vm.noContentAll());
+                },
+                hideNoItemsAll: function () {
+                    return (vm.stillWaiting() || !vm.noContentAll());
+                },
+                selectDetail: function selectDetail(questionId) {
+                    $state.go('questions_detail', {
+                        questionId: questionId,
+                        parentSectionId: vm.parentSection.id,
+                        parentSurveyId: vm.parentSurvey.id
                     });
-                } else if (question.referenceCount === 0) {
+                },
+                selectDetailGlobal: function (questionId) {
+                    $state.go('questions_detail', {
+                        questionId: questionId
+                    });
+                },
+                addQuestion: function () {
+                    $state.go('questions_add', {
+                        parentSectionId: vm.parentSection.id,
+                        parentSurveyId: vm.parentSurvey.id
+                    });
+                },
+                addQuestionGlobal: function () {
+                    $state.go('questions_addglobal');
+                },
+                addFromExisting: function () {
+                    questionsSrvc.isWaiting(true);
+                    $state.go('questions_addfe', {
+                        parentSectionId: vm.parentSection.id,
+                        parentSurveyId: vm.parentSurvey.id
+                    });
+                    questionsSrvc.updateRemainingQuestions().then(function () {
+                        $state.reload();
+                        questionsSrvc.isWaiting(false);
+                    });
+                },
+                showDeleteAlert: function ($event, questionId) {
+                    $event.stopPropagation();
+
+                    var question = questionsSrvc.getQuestionAt(questionId);
                     $ionicPopup.confirm({
                         title: 'Delete Question',
                         template: 'Are you sure you want to delete \'' + question.questionText + '\'?'
@@ -120,13 +97,44 @@
                         if (response) {
                             var removeIndex = vm.questions.findIndex(quest => quest.id === questionId);
                             vm.questions.splice(removeIndex, 1);
-                            questionsSrvc.deleteQuestion(question);
+                            vm.parentSection.questionIds.splice(removeIndex, 1);
+                            question.referenceCount = question.referenceCount - 1;
+                            questionsSrvc.updateQuestion(question).then(function () {
+                                sectionsSrvc.updateSection(vm.parentSection).then(function () {
+                                    sectionsSrvc.updateSections(vm.parentSurvey.sectionIds).then(function () {
+                                        questionsSrvc.updateQuestions(vm.parentSection.questionIds);
+                                    });
+                                });
+                            });
                         } else {
                             console.log('User pressed cancel');
                         }
                     });
+                },
+                showDeleteAlertGlobal: function ($event, questionId) {
+                    $event.stopPropagation();
+
+                    var question = questionsSrvc.getQuestionAtGlobal(questionId);
+                    if (question.referenceCount > 0) {
+                        $ionicPopup.alert({
+                            title: 'Can\'t delete question from global list!',
+                            template: 'Questions can only be deleted via the relevant section.'
+                        });
+                    } else if (question.referenceCount === 0) {
+                        $ionicPopup.confirm({
+                            title: 'Delete Question',
+                            template: 'Are you sure you want to delete \'' + question.questionText + '\'?'
+                        }).then(function (response) {
+                            if (response) {
+                                var removeIndex = vm.allQuestions.findIndex(quest => quest.id === questionId);
+                                vm.allQuestions.splice(removeIndex, 1);
+                                questionsSrvc.deleteQuestion(question);
+                            } else {
+                                console.log('User pressed cancel');
+                            }
+                        });
+                    }
                 }
-            }
            
         });
     }
